requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.1
cachelib==0.9.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.5
decorator==5.1.1
distro==1.9.0
elevenlabs==0.2.27
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.3
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.0
Flask-Session==0.8.0
flask-talisman==1.1.0
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==7.1.0
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.50
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.53
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
msgspec==0.18.6
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==0.28.0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
plotly==5.22.0
priority==2.0.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
Quart==0.19.5
regex==2024.4.28
requests==2.31.0
retrying==1.3.4
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.3
SQLAlchemy==2.0.29
srt==3.5.3
stack-data==0.6.3
starlette==0.37.2
taskgroup==0.0.0a4
tenacity==8.2.3
tiktoken==0.6.0
tomli==2.0.1
tqdm==4.66.4
traitlets==5.14.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vosk==0.3.45
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.0
